Move test sample to test samples location
