+++ java-checks-test-sources/default/src/test/java/checks/tests/AssertThrowsInsteadOfTryCatchFailCheckSample.java
@@ -1,4 +1,4 @@
-package checks;
+package checks.tests;
 
 import org.junit.jupiter.api.Test;
